[FEATURE] [SECURITY] cookies 에 access 토큰(개인 정보 포함) 발급

+++ src/main/java/com/soop/jwtsecurity/entityDTO/UserEntity.java
@@ -18,7 +18,7 @@ public class UserEntity {
 //    private String role;
 //    private String test;
 
-    private Long userCode;
+    private int userCode;
     private String nickName;
     private String email;
     private String gender;
